refactor: `models/entity` 폴더, 내부 파일 이름 변경
