+++ src/models/entities/index.ts
@@ -1,0 +1,4 @@
+export * from './User.entity';
+export * from './SubmitCode.entity';
+export * from './Problem.entity';
+export * from './TestCase.entity';
