--- src/models/entity/index.ts
@@ -1,4 +1,0 @@
-export * from './User';
-export * from './SubmitCode';
-export * from './Problem';
-export * from './TestCase';
