+++ src/modules/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { User } from '@/models/entity';
+import { User } from '@/models/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
