+++ src/modules/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from '@/models/entity';
+import { User } from '@/models/entities';
 import { Repository } from 'typeorm';
 import { OAuthApp } from 'octokit';
 import { configService } from '@/config/config.service';
