+++ src/modules/problem/problem.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Problem, SubmitCode, TestCase, User } from '@/models/entity';
+import { Problem, SubmitCode, TestCase, User } from '@/models/entities';
 import { Repository } from 'typeorm';
 import { judge } from '@/utils';
 import { JudgeStatus } from './dto/judgeStatus.dto';
