+++ src/modules/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { User } from '@/models/entity';
+import { User } from '@/models/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
